+++ lib/callable_tree/node/internal/strategy/broadcast.rb
@@ -5,6 +5,8 @@ module CallableTree
     module Internal
       module Strategy
         class Broadcast
+          include Strategy
+
           def call(nodes, *inputs, **options)
             nodes.map do |node|
               node.call(*inputs, **options) if node.match?(*inputs, **options)