+++ spec/callable_tree/node/internal_spec.rb
@@ -584,12 +584,24 @@ RSpec.describe CallableTree::Node::Internal do
     end
   end
 
-  shared_context 'with parent node' do
-    let!(:parent_node) do
-      CallableTree::Node::Root.new.tap do |root_node|
-        root_node.children << node
-        node.send(:parent=, root_node)
+  describe '#seek?' do
+    subject { node.seek? }
+
+    let(:node) { ::Class.new { include CallableTree::Node::Internal }.new }
+
+    context 'when strategy is `seek`' do
+      before { node.send(:strategy=, described_class::Strategy::Seek.new) }
+      it { is_expected.to be true }
+    end
+
+    context 'when strategy is not `seek`' do
+      before do
+        node.send(:strategy=, [
+          described_class::Strategy::Broadcast.new,
+          described_class::Strategy::Compose.new
+        ].sample)
       end
+      it { is_expected.to be false }
     end
   end
 
@@ -608,15 +620,6 @@ RSpec.describe CallableTree::Node::Internal do
 
       it { is_expected.not_to be node }
       it { expect(subject.send(:strategy)).to be_a described_class::Strategy::Seek }
-
-      context 'when node has parent' do
-        include_context 'with parent node'
-        it { expect(subject.root?).to be true }
-      end
-
-      context 'when node does not have parent' do
-        it { expect(subject.root?).to be true }
-      end
     end
 
     context 'when current strategy is `compose`' do
@@ -624,15 +627,6 @@ RSpec.describe CallableTree::Node::Internal do
 
       it { is_expected.not_to be node }
       it { expect(subject.send(:strategy)).to be_a described_class::Strategy::Seek }
-
-      context 'when node has parent' do
-        include_context 'with parent node'
-        it { expect(subject.root?).to be true }
-      end
-
-      context 'when node does not have parent' do
-        it { expect(subject.root?).to be true }
-      end
     end
   end
 
@@ -661,6 +655,27 @@ RSpec.describe CallableTree::Node::Internal do
     end
   end
 
+  describe '#broadcast?' do
+    subject { node.broadcast? }
+
+    let(:node) { ::Class.new { include CallableTree::Node::Internal }.new }
+
+    context 'when strategy is `broadcast`' do
+      before { node.send(:strategy=, described_class::Strategy::Broadcast.new) }
+      it { is_expected.to be true }
+    end
+
+    context 'when strategy is not `broadcast`' do
+      before do
+        node.send(:strategy=, [
+          described_class::Strategy::Seek.new,
+          described_class::Strategy::Compose.new
+        ].sample)
+      end
+      it { is_expected.to be false }
+    end
+  end
+
   describe '#broadcast' do
     subject { node.broadcast }
 
@@ -669,15 +684,6 @@ RSpec.describe CallableTree::Node::Internal do
     context 'when current strategy is `seek`' do
       it { is_expected.not_to be node }
       it { expect(subject.send(:strategy)).to be_a described_class::Strategy::Broadcast }
-
-      context 'when node has parent' do
-        include_context 'with parent node'
-        it { expect(subject.root?).to be true }
-      end
-
-      context 'when node does not have parent' do
-        it { expect(subject.root?).to be true }
-      end
     end
 
     context 'when current strategy is `broadcast`' do
@@ -692,15 +698,6 @@ RSpec.describe CallableTree::Node::Internal do
 
       it { is_expected.not_to be node }
       it { expect(subject.send(:strategy)).to be_a described_class::Strategy::Broadcast }
-
-      context 'when node has parent' do
-        include_context 'with parent node'
-        it { expect(subject.root?).to be true }
-      end
-
-      context 'when node does not have parent' do
-        it { expect(subject.root?).to be true }
-      end
     end
   end
 
@@ -729,6 +726,27 @@ RSpec.describe CallableTree::Node::Internal do
     end
   end
 
+  describe '#compose?' do
+    subject { node.compose? }
+
+    let(:node) { ::Class.new { include CallableTree::Node::Internal }.new }
+
+    context 'when strategy is `compose`' do
+      before { node.send(:strategy=, described_class::Strategy::Compose.new) }
+      it { is_expected.to be true }
+    end
+
+    context 'when strategy is not `compose`' do
+      before do
+        node.send(:strategy=, [
+          described_class::Strategy::Seek.new,
+          described_class::Strategy::Broadcast.new,
+        ].sample)
+      end
+      it { is_expected.to be false }
+    end
+  end
+
   describe '#compose' do
     subject { node.compose }
 
@@ -737,15 +755,6 @@ RSpec.describe CallableTree::Node::Internal do
     context 'when current strategy is `seek`' do
       it { is_expected.not_to be node }
       it { expect(subject.send(:strategy)).to be_a described_class::Strategy::Compose }
-
-      context 'when node has parent' do
-        include_context 'with parent node'
-        it { expect(subject.root?).to be true }
-      end
-
-      context 'when node does not have parent' do
-        it { expect(subject.root?).to be true }
-      end
     end
 
     context 'when current strategy is `broadcast`' do
@@ -753,15 +762,6 @@ RSpec.describe CallableTree::Node::Internal do
 
       it { is_expected.not_to be node }
       it { expect(subject.send(:strategy)).to be_a described_class::Strategy::Compose }
-
-      context 'when node has parent' do
-        include_context 'with parent node'
-        it { expect(subject.root?).to be true }
-      end
-
-      context 'when node does not have parent' do
-        it { expect(subject.root?).to be true }
-      end
     end
 
     context 'when current strategy is `compose`' do
@@ -837,5 +837,56 @@ RSpec.describe CallableTree::Node::Internal do
     end
 
     it { is_expected.to eq result }
+  end
+
+  describe '#clone' do
+    subject { node.clone }
+
+    let(:tree) do
+      CallableTree::Node::Root.new.append!(
+        IdNode.new(:a).append!(
+          IdLeaf.new(:b)
+        )
+      )
+    end
+
+    context 'of root_node' do
+      let(:node) { tree }
+      it { is_expected.not_to be tree }
+
+      it 'should have cloned child nodes' do
+        expect(subject[0]).not_to be node[0]
+      end
+
+      it 'should be linked from child node as parent node' do
+        expect(subject[0].parent).to be subject
+      end
+    end
+
+    context 'of a_node' do
+      let(:node) { tree[0] }
+      it { is_expected.not_to be tree[0] }
+
+      it 'should have not parent node' do
+        expect(subject.parent).to be nil
+      end
+
+      it 'should have cloned child nodes' do
+        expect(subject[0]).not_to be node[0]
+      end
+
+      it 'should be linked from child node as parent node' do
+        expect(subject[0].parent).to be subject
+      end
+    end
+
+    context 'of b_node' do
+      let(:node) { tree[0][0] }
+      it { is_expected.not_to be tree[0][0] }
+
+      it 'should have not parent node' do
+        expect(subject.parent).to be nil
+      end
+    end
   end
 end